+++ l2/agentDefsL2/agentDefsL2.ts
@@ -16,6 +16,8 @@ import '/_102020_/l2/agentDefsL2/steps/input20/agentD2Input.js';
 import '/_102020_/l2/agentDefsL2/steps/contracts30/agentD2Contracts.js';
 import '/_102020_/l2/agentDefsL2/steps/shared40/agentD2Shared.js';
 import '/_102020_/l2/agentDefsL2/steps/shared40/agentD2SharedPage.js';
+import '/_102020_/l2/agentDefsL2/steps/pages50/agentD2Pages.js';
+import '/_102020_/l2/agentDefsL2/steps/pages50/agentD2PagesPage.js';
 
 export function createAgent(): IAgentAsync {
   return {